requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioredis==2.0.1
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apify_client==1.9.1
apify_shared==1.2.1
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
captcha==0.6.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
docarray==0.40.0
docstring_parser==0.16
email_validator==2.2.0
extra-streamlit-components==0.1.71
fastapi==0.115.7
fastapi-admin==1.0.4
filetype==1.2.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iso8601==2.1.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-google-vertexai==2.0.11
langchain-openai==0.3.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.5
langsmith==0.3.1
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.23.0
numpy==2.2.2
openai==1.60.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pendulum==3.0.0
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2-binary==2.9.10
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.2.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sentry-sdk==2.20.0
setuptools==75.1.0
shapely==2.0.6
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqladmin==0.20.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.2
stqdm==0.0.5
streamlit==1.41.1
streamlit-authenticator==0.4.1
tenacity==9.0.0
tiktoken==0.8.0
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
tortoise-orm==0.24.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.24
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==14.2
wheel==0.44.0
WTForms==3.1.2
yarl==1.18.3
zstandard==0.23.0
